requirements.txt
selenium==3.141.0
pyYAML==5.4.1